# api/generated/openapi.yaml
...
openapi: 3.1.0
info:
  title: AsiaTravelPlan Mobile API
  version: 2026-03-02.1
  summary: Contract for the in-house iPhone app and the AsiaTravelPlan backend.
  description: |-
    This contract is generated from the CUE abstract model (model/api, model/entities, model/enums, model/common).
    It describes the request and response payloads the iPhone app and frontend may rely on.

    Important:
    - The mobile app must not depend on backend storage details.
    - Breaking changes should increment the contract version and normally require a mobile app update.
    - Only these app roles are supported: atp_admin, atp_manager, atp_accountant, atp_staff, atp_tour_editor.
servers:
- url: https://api-staging.asiatravelplan.com
  description: Staging API
- url: http://localhost:8787
  description: Local development API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LanguageCode:
      type: string
      enum:
      - en
      - ar
      - fr
      - zh
      - ja
      - ko
      - vi
      - ms
      - de
      - es
      - it
      - ru
      - nl
      - pl
      - da
      - sv
      - 'no'
      x-enum-catalog:
      - code: en
      - code: ar
      - code: fr
      - code: zh
      - code: ja
      - code: ko
      - code: vi
      - code: ms
      - code: de
      - code: es
      - code: it
      - code: ru
      - code: nl
      - code: pl
      - code: da
      - code: sv
      - code: 'no'
      x-enum-options:
      - value: en
        label: en
      - value: ar
        label: ar
      - value: fr
        label: fr
      - value: zh
        label: zh
      - value: ja
        label: ja
      - value: ko
        label: ko
      - value: vi
        label: vi
      - value: ms
        label: ms
      - value: de
        label: de
      - value: es
        label: es
      - value: it
        label: it
      - value: ru
        label: ru
      - value: nl
        label: nl
      - value: pl
        label: pl
      - value: da
        label: da
      - value: sv
        label: sv
      - value: 'no'
        label: 'no'
...
    BookingSourceChannel:
      type: string
      enum:
      - other
      - website
      - email
      - whatsapp
      - facebook_messenger
      - google_maps
      - instagram
      - phone_call
      - walk_in
      - zalo
      x-enum-catalog:
      - code: other
      - code: website
      - code: email
      - code: whatsapp
      - code: facebook_messenger
      - code: google_maps
      - code: instagram
      - code: phone_call
      - code: walk_in
      - code: zalo
      x-enum-options:
      - value: other
        label: other
      - value: website
        label: website
      - value: email
        label: email
      - value: whatsapp
        label: whatsapp
      - value: facebook_messenger
        label: facebook_messenger
      - value: google_maps
        label: google_maps
      - value: instagram
        label: instagram
      - value: phone_call
        label: phone_call
      - value: walk_in
        label: walk_in
      - value: zalo
        label: zalo
    BookingReferralKind:
      type: string
      enum:
      - none
      - other_customer
      - b2b_partner
      - atp_staff
      x-enum-catalog:
      - code: none
      - code: other_customer
      - code: b2b_partner
      - code: atp_staff
      x-enum-options:
      - value: none
        label: none
      - value: other_customer
        label: other_customer
      - value: b2b_partner
        label: b2b_partner
      - value: atp_staff
        label: atp_staff
...
    AtpStaffProfile:
      type: object
      required:
      - username
      properties:
        username:
          type: string
        name:
          type: string
          nullable: true
        full_name:
          type: string
          nullable: true
        friendly_short_name:
          type: string
          nullable: true
        team_order:
          type: integer
          nullable: true
        picture_ref:
          type: string
          nullable: true
        languages:
          type: array
          items:
            "$ref": "#/components/schemas/LanguageCode"
        destinations:
          type: array
          items:
            "$ref": "#/components/schemas/CountryCode"
...
    BookingTravelPlanAttachment:
      type: object
      required:
      - id
      - filename
      - storage_path
      - page_count
      - sort_order
      properties:
        id:
          type: string
        filename:
          type: string
        storage_path:
          type: string
        page_count:
          type: integer
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
          nullable: true
    BookingTravelPlan:
      type: object
      required: []
      properties:
        destinations:
          type: array
          items:
            "$ref": "#/components/schemas/CountryCode"
        days:
          type: array
          items:
            "$ref": "#/components/schemas/BookingTravelPlanDay"
        attachments:
          type: array
          items:
            "$ref": "#/components/schemas/BookingTravelPlanAttachment"
...
    BookingTravelPlanPdfReadModel:
      type: object
      required:
      - id
      - filename
      - page_count
      - created_at
      - sent_to_customer
      - pdf_url
      properties:
        id:
          type: string
        filename:
          type: string
        page_count:
          type: integer
        created_at:
          type: string
          format: date-time
        sent_to_customer:
          type: boolean
        comment:
          type: string
          nullable: true
        pdf_url:
          type: string
...
    BookingReadModel:
      type: object
      required:
      - id
      - offer
      - travel_plan_translation_status
      - generated_offer_email_enabled
      - translation_enabled
      - created_at
      - updated_at
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        assigned_keycloak_user_id:
          type: string
          nullable: true
        source_channel:
          "$ref": "#/components/schemas/BookingSourceChannel"
          nullable: true
        referral_kind:
          "$ref": "#/components/schemas/BookingReferralKind"
          nullable: true
        referral_label:
          type: string
          nullable: true
        referral_staff_user_id:
          type: string
          nullable: true
        assigned_keycloak_user_label:
          type: string
          nullable: true
        assigned_atp_staff:
          "$ref": "#/components/schemas/AtpStaffProfile"
          nullable: true
        core_revision:
          type: integer
          nullable: true
        notes_revision:
          type: integer
          nullable: true
        persons_revision:
          type: integer
          nullable: true
        travel_plan_revision:
          type: integer
          nullable: true
        offer_revision:
          type: integer
          nullable: true
        payment_documents_revision:
          type: integer
          nullable: true
        travel_styles:
          type: array
          items:
            type: string
        travel_start_day:
          type: string
          format: date
          nullable: true
        travel_end_day:
          type: string
          format: date
          nullable: true
        number_of_travelers:
          type: integer
          nullable: true
        preferred_currency:
          "$ref": "#/components/schemas/ATPCurrencyCode"
          nullable: true
        customer_language:
          "$ref": "#/components/schemas/LanguageCode"
          nullable: true
        notes:
          type: string
          nullable: true
        persons:
          type: array
          items:
            "$ref": "#/components/schemas/BookingPerson"
        travel_plan:
          "$ref": "#/components/schemas/BookingTravelPlan"
          nullable: true
        web_form_submission:
          "$ref": "#/components/schemas/BookingWebFormSubmission"
          nullable: true
        offer:
          "$ref": "#/components/schemas/BookingOfferReadModel"
        travel_plan_pdfs:
          type: array
          items:
            "$ref": "#/components/schemas/BookingTravelPlanPdfReadModel"
        generated_offers:
          type: array
          items:
            "$ref": "#/components/schemas/GeneratedBookingOfferReadModel"
        travel_plan_translation_status:
          "$ref": "#/components/schemas/TranslationStatusSummary"
        generated_offer_email_enabled:
          type: boolean
        translation_enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BookingListFilters:
      type: object
      required: []
      properties:
        assigned_keycloak_user_id:
          type: string
          nullable: true
        search:
          type: string
          nullable: true
    BookingList:
      type: object
      required:
      - pagination
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/BookingReadModel"
        pagination:
          "$ref": "#/components/schemas/Pagination"
        filters:
          "$ref": "#/components/schemas/BookingListFilters"
          nullable: true
        sort:
          type: string
          nullable: true
...
    BookingDetail:
      type: object
      required:
      - booking
      properties:
        booking:
          "$ref": "#/components/schemas/BookingReadModel"
        unchanged:
          type: boolean
          nullable: true
...
    Pagination:
      type: object
      required:
      - page
      - page_size
      - total_items
      - total_pages
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
...
    BookingNameUpdateRequest:
      type: object
      required: []
      properties:
        expected_core_revision:
          type: integer
          nullable: true
        name:
          type: string
          nullable: true
        actor:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
        detail:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        request_id:
          type: string
          nullable: true
security:
- bearerAuth: []
paths:
...
  "/api/v1/bookings":
    get:
      operationId: bookings
      summary: Bookings
      tags:
      - Bookings
      security:
      - bearerAuth: []
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BookingList"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
    post:
      operationId: booking_create
      summary: Booking Create
      tags:
      - Bookings
      security:
      - bearerAuth: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/BookingCreateRequest"
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BookingDetail"
        '201':
          description: Created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BookingDetail"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
...
  "/api/v1/bookings/{booking_id}/name":
    patch:
      operationId: booking_name
      summary: Booking Name
      tags:
      - Bookings
      security:
      - bearerAuth: []
      parameters:
      - name: booking_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/BookingNameUpdateRequest"
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BookingDetail"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
...